Add lib.rs
